+++ src/main/java/io/github/mmpodkanski/computershop/order/dto/OrderItemDto.java
@@ -2,11 +2,13 @@ package io.github.mmpodkanski.computershop.order.dto;
 
 import io.github.mmpodkanski.computershop.product.dto.ProductDto;
 
+import java.math.BigDecimal;
+
 public interface OrderItemDto {
     static OrderItemDto create(
             int id,
             int quantity,
-            double price,
+            BigDecimal price,
             ProductDto product
     ) {
         return new OrderItemDtoImpl(id, quantity, price, product);
@@ -16,17 +18,17 @@ public interface OrderItemDto {
 
     int getQuantity();
 
-    double getPrice();
+    BigDecimal getPrice();
 
     ProductDto getProduct();
 
     class OrderItemDtoImpl implements OrderItemDto {
         private final int id;
         private final int quantity;
-        private final double price;
+        private final BigDecimal price;
         private final ProductDto product;
 
-        OrderItemDtoImpl(final int id, final int quantity, final double price, final ProductDto product) {
+        OrderItemDtoImpl(final int id, final int quantity, final BigDecimal price, final ProductDto product) {
             this.id = id;
             this.quantity = quantity;
             this.price = price;
@@ -44,7 +46,7 @@ public interface OrderItemDto {
         }
 
         @Override
-        public double getPrice() {
+        public BigDecimal getPrice() {
             return price;
         }
